+++ plugins/tasks/throw/src/main/java/com/walmartlabs/concord/plugins/throwex/ThrowExceptionTask.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.project.yaml.converter;
+package com.walmartlabs.concord.plugins.throwex;
 
 /*-
  * *****
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.project.yaml.converter;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,20 +20,18 @@ package com.walmartlabs.concord.project.yaml.converter;
  * =====
  */
 
-import com.walmartlabs.concord.project.yaml.YamlConverterException;
-import com.walmartlabs.concord.project.yaml.model.YamlExit;
-import io.takari.bpm.model.TerminateEvent;
+import com.walmartlabs.concord.sdk.Task;
 
-public class YamlExitConverter implements StepConverter<YamlExit> {
+import javax.inject.Named;
 
-    @Override
-    public Chunk convert(ConverterContext ctx, YamlExit s) throws YamlConverterException {
-        Chunk c = new Chunk();
+@Named("throw")
+public class ThrowExceptionTask implements Task {
 
-        String id = ctx.nextId();
-        c.addElement(new TerminateEvent(id));
-        c.addSourceMap(id, toSourceMap(s, "Exit"));
-
-        return c;
+    public void call(Object o) throws Exception {
+        if (o instanceof Exception) {
+            throw (Exception) o;
+        } else {
+            throw new RuntimeException(o != null ? o.toString() : "n/a");
+        }
     }
 }